Center bar module-right now dynamic
With my current setup, my laptop monitor acts as my left monitor of my
three monitor setup. To simplify things it will remain the center bar.
The right modules in this setup need to be slightly modified. On a one
monitor setup, my nm-applet in the tray is on the one screen, so network
modules are not needed. On a three monitor setup, the tray is on my
center screen (left external monitor). In this setup I prefer to have
the network modules on my laptop bar.