Fix logic handling for polybar launch
The way the for loop was handling the monitors was really inneficient,
Pass through array instead of a command result directly and replaced
spaces with underscores to simplify the array creation. This garauntees
an array of 3 items instead of around 40. Also fixed the
vertical/horizontal handling for the right bar, it was never working in
the first place.